Fix collision detection
Fix collision detection code to check against a a box rather than a point.

This ensures that the player never ends too close to the edge of a cell which
causes clipping of the forward facing quads.  Also implement movement clipping
so that the player is able to move if it's colliding just on one of the axes.

Signed-off-by: Pekka Enberg <penberg@iki.fi>